task_9/requirements.txt
# No external dependencies required for the basic version
# For AI integration, add: openai 

openai
python-dotenv 